# Assign id attributes first in Active Record attribute assignment