### Motivation / Background

This Pull Request has been created because store assignment behaviour changed subtly since https://github.com/rails/rails/pull/52792:

```rb
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", github: "rails/rails"

  gem "sqlite3"# , "~> 1.4"
  gem "benchmark-ips"
end

require "active_record"
require "minitest/autorun"
require "logger"

# This connection will do for database-independent bug reports.
ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(STDOUT)

ActiveRecord::Schema.define do
  create_table :posts, force: true do |t|
    t.text :settings
  end
end

class Post < ActiveRecord::Base
  store :settings, accessors: [:a]
end

class BugTest < Minitest::Test
  def test_association_stuff
    post = Post.create!

    post.update!(settings: { a: "a" }, a: "b")

    assert_equal("b", post.a)
  end
end
```
(fails on main, passes on 7.2, and then fails on 7.1)

Instead of assigning hashes later, we can assign IDs first to maintain the same behaviour for nested attribute assignment. Deferring hash assignment causes problems in store scenarios where an attribute can be assigned two different ways, which had worked since Rails 7.2.

### Detail

This Pull Request changes attribute assignment to assign foreign keys first.

### Additional information

```rb
require "benchmark/ips"

post = Post.create!

Benchmark.ips do |x|
  x.report(:assignment) { post.assign_attributes(settings: { a: "a" }, a: "b") }
end
```

Before:
```
ruby 3.3.4 (2024-07-09 revision be1089c8ec) [arm64-darwin23]
Warming up --------------------------------------
          assignment     2.387k i/100ms
Calculating -------------------------------------
          assignment     23.466k (± 4.6%) i/s   (42.61 μs/i) -    119.350k in   5.097964s
```

After:
```
ruby 3.3.4 (2024-07-09 revision be1089c8ec) [arm64-darwin23]
Warming up --------------------------------------
          assignment     2.358k i/100ms
Calculating -------------------------------------
          assignment     22.578k (± 5.0%) i/s   (44.29 μs/i) -    113.184k in   5.026557s
```
Potentially a little slower, but not significantly.


### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.